+++ brewtils/decorators.py
@@ -246,7 +246,7 @@ def parameter(
     display_name=None,  # type: Optional[str]
     optional=None,  # type: Optional[bool]
     default=None,  # type: Optional[Any]
-    filename=None,  # type: Optional[str]
+    file_name=None,  # type: Optional[str]
     description=None,  # type: Optional[str]
     choices=None,  # type: Optional[Union[Callable, Dict, Iterable, str]]
     parameters=None,  # type: Optional[List[Parameter]]
@@ -285,7 +285,7 @@ def parameter(
         optional: Boolean indicating if this parameter must be specified.
         default: The value this parameter will be assigned if not overridden when
             creating a request.
-        filename: Default string value for file input.
+        file_name: Default string value for file input.
         description: An additional string that will be displayed in the user interface.
         choices: List or dictionary specifying allowed values. See documentation for
             more information.
@@ -333,7 +333,7 @@ def parameter(
             display_name=display_name,
             optional=optional,
             default=default,
-            filename=filename,
+            file_name=file_name,
             description=description,
             choices=choices,
             parameters=parameters,
@@ -354,7 +354,7 @@ def parameter(
         display_name=display_name,
         optional=optional,
         default=default,
-        filename=filename,
+        file_name=file_name,
         description=description,
         choices=choices,
         parameters=parameters,
@@ -889,7 +889,7 @@ def _initialize_parameter(
     display_name=None,
     optional=None,
     default=None,
-    filename=None,
+    file_name=None,
     description=None,
     choices=None,
     parameters=None,
@@ -931,7 +931,7 @@ def _initialize_parameter(
         display_name=display_name,
         optional=optional,
         default=default,
-        filename=filename,
+        file_name=file_name,
         description=description,
         choices=choices,
         parameters=parameters,